requirements.txt
Flask>=2.0.1
scikit-learn>=1.0.2
pandas>=1.3.3
numpy>=1.21.2
joblib>=1.1.0
Flask-WTF>=0.15.1
Werkzeug>=2.0.1
python-dotenv>=0.19.0
setuptools>=65.5.1
gunicorn>=20.1.0
pytest>=7.0.0
flake8>=4.0.1 